S13P31A703 fix:is_delect 제외하게 추가

+++ backend/runnity/src/main/java/com/runnity/history/dto/response/StatsSummaryResponse.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record StatsSummaryResponse(
         Float totalDistance,           // 총 거리 (km)
         Integer totalTime,             // 총 운동 시간 (초)
-        Float avgPace,                 // 평균 페이스 (초/km)
+        Integer avgPace,                 // 평균 페이스 (초/km)
         Integer totalRunDays,          // 달린 날 수
         List<PeriodStatResponse> periodStats,  // 구간별 달린 거리
         List<RunRecordResponse> personals,   // 개인 러닝 최근 5개
